Add files via upload
this file takes in a keyword argument as well as file arguments. once the keyword is found in a line of the file, it is printed. Lines do not print if no keyword is found, and there are error messages for an invalid file as well as no argument on the command line.